fix: component EditViewerInfo visualisation, upload avatar/data logic, LightPopup
