+++ src/features/edit-viewer-info/types.ts
@@ -2,6 +2,6 @@ export interface ViewerInputData {
   target: {
     value: string;
     name: string;
-    files: any;
+    files: FileList | null;
   };
 }
